Fixed pagewidth for commit messages
My default pagewidth is 80 characters, but recently I've noticed that my
git commit messages tend to look bad and break in the wrong places. I
found that the recommended pagewidth for commit messages is 72
characters, so I set it to change to that width only when I'm editing
commit messages.