Updated emacs config
Implemented system that keeps temporary buffers from being switched to
automatically. The list of regexes which is used to tell helm which
buffers to hide is now also used to rearrange the buffer list. A
function is run every time the buffer list changes, and this function
will arrange the list in the following order: buffers not matched by the
regexes, the scratch buffer, buffers matched by the regexes. The
ordering within each of the two groups is preserved. This means that
when a window switches to whatever the next buffer happens to be after a
kill or bury, the chosen buffer will not be a temporary buffer, and will
only be the scratch buffer if no other non-temporary buffers exist (it
will also create a scratch buffer if none exists).