Updated emacs config
- Reimplemented buffer hiding so it causes less weirdness
- Added pretty modelines
- Fixed tab completion popwin issue
- Simplified and fixed up autoroot advice
- Rewrote some things to make them look nicer
- Rearranged some things to be more reasonable

Buffer hiding works completely differently now. It used to be that each
time the buffer list changed, it was rearranged to put the hidden
buffers at the bottom of the stack. This worked alright, but there were
some issues (hidden buffers would display when the only non-hidden
buffer is *scratch* and it gets buried, buffers visible in other windows
would not have less priority, next-buffer and friends would still show
hidden buffers, etc). However, emacs has buffer hiding built in, and
this new implementation takes advantage of that. Emacs hides any buffer
that starts with a space, anywhere that this is sensible. The new
implementation advises all the functions that check if a buffer starts
with a space, and redefines buffer-name in those functions to prepend a
space to the buffer name if it matches the boring buffer regex list.
Now, just those functions think these buffers start with a space, and
they hide them appropriately.