Updated emacs config
- greatly improved new side window manager
- added scala mode
- added advice to get Evil to properly set initial state automatically
- removed noselect from completion buffer and magit process buffer (was
  interfering with new side window management)
- improved term-mode message handling
- put evil subword code in an init hook so reloading init file works
- improved state display in modeline and other minor things
- removed useless hash characters denoting function symbols
- made ,x print to buffer in all lisp-interaction-mode rather than just
  the scratch buffer
- made ,z print a status whenever it's used

The side window manager has been greatly improved, because a side window
tracking system has been added. The tracking system stores a persistent
tree structure as a frame parameter in each frame, and the tree is
modified as side windows are created and destroyed. This allows the
proper windows to be killed when switching or deleting windows, and it
allows proper recursive behavior when opening and closing side window
buffers.